Added more cutoffs and intervals
So far it seems that the algorithm for the cutoffs is always going to be every other number in the fibonacci sequence, and the intervals will always be the same as the last interval set, except the smallest of the set needs to be reevaluated. I'm still working out what that algorithm needs to be, but it's clear that an interval with start as a prime, and then  must be multiplied by the number of resulting spirals to get the final interval (il.e. if 5 creates 11 accurate spirals, then the resulting interval is 5 * 11 = 55)