Fix comments that incorrectly talk about golden ratio instead of golden angle
I've been accidentally referring to the golden angle as the golden ratio this whole time. To be clear, a fibonacci sphere is generated using the golden angle which is pi * (3 - sqrt(5)). while I was calling it the golden ratio, the actual formula used was the golden angle, so only the comments needed to be updated.